refactor: LineDao 를 인터페이스로 분리
