+++ src/main/java/wooteco/subway/domain/Line.java
@@ -3,8 +3,8 @@ package wooteco.subway.domain;
 public class Line {
 
     private Long id;
-    private String name;
-    private String color;
+    private final String name;
+    private final String color;
 
     public Line(Long id, String name, String color) {
         this.id = id;
